Allow to use reducers and tables on the client itself
